TXT: join text records
TXT records that are larger then 255 characters are split into chunks.
This means that the result of a TXT record is always
Array[Array[String]].

This 355 character limit is more a detail about DNS which is not very
useful when presenting data to higher interfaces. as such join the inner
elements so we have a more normalise Array[String] return_type

This is a backwards incompatible breaking change