138: Allow nil value for config_info
We want to be able to pass an explicit nil value for the config_info
parameter so that we can allow passing through that value from higher
level functions e.g. lookup.

The author originally thought that optional_param would allow this
behaviour however that just allows for different signature of the
method and doesn't effect the type validation.  as such we need to also
make the parameter type Optional[T] so that we all passing explicit nil
values.

see: https://tickets.puppetlabs.com/browse/PUP-11066
Fixes #131